kitty: Switch to Monaspace from Dank Mono
Dank Mono contains non-fixed-width glyphs and that makes macOS (and
probably other OSes) think it's not a fixed-width font. With kitty's
recent update, it only let us choose a font that recognized as
fixed-width by system, so kitty does not render Dank Mono anymore.
It's breaking change and so annoying, but kitty is still in v0.x.

Honestly, Monaspace is much legible than Dank Mono...