nvim: Fix incorrect indent inside block comment in JS/TS
<https://github.com./nvim-treesitter/nvim-treesitter/issues/2544>

Even though the issue is closed as completed, the problem still occurs.
An issue on another repository back-linked in the above issue includes a
comment that enabling nvim-treesitter's `indent` option solves this.
This option may cause more problems than the benefits it brings, but
it's working at least for now.