kitty: Show window decorations
Kitty sometimes opens a window on a non-primary display for some
reasons (probably a bug). In that occasion, it's quite annoying to move
the window to primary display in Mission Control then maximize the
window via menu bar.

macOS window decoration on terminal is extremely ugly, but it's still
better than no control.