ghostty: Disable discretionary ligatures
I noticed a word "ページ" is rendered as "㌻  ".

Suspected ligature settings, removed "calt" font-feature, which
activates Monaspace's Texture headling feature. Still "㌻  ".

After searching Ghostty issue tracker, I found a culprit:

<https://github.com/ghostty-org/ghostty/issues/1710>

It was "dlig" font-feature, which is not recommended to be on by default
because of its nature. Ghostty seems to set to on by default, for some
reasons.

It completely spoils Japanese text rendering and brings no benefits.

The linked issue has comments with insightful links.