jj: Prevent pushing commits prefixed with "[WIP]"
"wip:" prefix, which is used as an example in JJ docs, is not suitable
for many projects because "scope:" is quite popular prefix in the wild.

Using that would leads to, for example, "wip: jj: Change wording".
This is confusing and does not stand out among other commits. WIP marker
should be easily spottable.

I tried to use "regex-" kind, but it's not available in this version of
JJ. I can't upgrade nixpkgs due to Ghostty is marked as broken in newer
nixpkgs.