License non-programs 0BSD and media files CC BY-ND 4.0
Zero-Clause BSD has much shorter license text that I can put the whole
text into comment header. It also allows non-attribution so anyone can
safely copy and paste whole code or a chunk of.

OSS licenses are made for software programs and do not work well with
media assets. Creative Commons is better choice for that purpose.
I choose BY-ND 4.0 because currently a screenshot file is the only media
file in this repo. If someone fork or reuse the code from this repo and
put a screenshot of their derivative work, it's made from scratch not
modifying the existing screenshot, because it is, uh, screenshot.