Update nixpkgs to latest unstable
To fix Syncthing failing to launch.

I've once upgraded nixpkgs to the latest version two days ago *locally*.
It was to test if the new Atuin package (which addressed Fish breaking
change) but it wasn't there, so I `jj restore`-ed. The problem is,
Syncthing migrated its database to the latest version and their
migration does not consider downgrading at all. Since the downgrade,
Syncthing fails to start with a migration error (an SQL error containing
`CREATE INDEX IF NOT EXISTS files_name_only ON files (name)`.)